Add files via upload
Text generator from the following task: You will create a program that can predict the next word in a pseudo-sentence based on the previous words in the sequence and the data that is used to create a statistical model. You will get a deeper understanding of natural language processing and preprocessing with the NLTK library, string operations, and the application of statistics in your code.