Add files via upload
In this project, I simulated the simple hacking process of a non-protected website (more specifically it was about getting admins right). I was a 5-step project with different conditions and ways of coming up with the correct password. On the final stage, the situation was following: the admin has missed the fact that guessing the first letter of the password (with the correct login) will give an Exception Error" message instead of the "Wrong password" one. But here's the thing: the admin probably just caught this exception. Catching an exception takes the computer a long time, so I use it to hack the system by counting the time period in which the response comes and finding out which starting symbols work out for the password.