Add files via upload
Data analysis project. Creating a report for the HR department. The project was consisting of 5 steps where each one was dedicated to formatting data. The final output can was got from the following task:
The first pivot table displays departments as rows, employees' current status, and their salary level as columns. The values should be the median number of monthly hours employees have worked. In the table, the HR boss wants to see only those departments where either one is true:
For the currently employed: the median value of the working hours of high-salary employees is smaller than the hours of the medium-salary employees, OR:
For the employees who left: the median value of working hours of low-salary employees is smaller than the hours of high-salary employees
The second pivot table is where each row is an employee's time in the company; the columns indicate whether an employee has had any promotion. The values are the last evaluation score's minimum, maximum, mean, and satisfaction level. Filter the table by the following rule: select only those rows where the previous mean evaluation score is higher for those without promotion than those who had.